Update docs from develop branch
The new documentation, though incomplete, is far better than what I had
before. However, I removed the references to cloning penguins to get the
test data, as it will not be available in the master branch (I didn't
pull the tests over from the develop branch, and don't intend to do so
until v0.2 is released).